\name{ssJIVE-package}
\alias{ssJIVE-package}
\alias{ssJIVE}
\docType{package}
\title{
\packageTitle{ssJIVE}
}
\description{
\packageDescription{ssJIVE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ssJIVE}
\packageIndices{ssJIVE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ssJIVE}

Maintainer: \packageMaintainer{ssJIVE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
